+++ packages/atomic-react/src/components/search/components.ts
@@ -1,4 +1,7 @@
-import {AtomicComponentError as LitAtomicComponentError} from '@coveo/atomic/components';
+import {
+  AtomicComponentError as LitAtomicComponentError,
+  AtomicIcon as LitAtomicIcon,
+} from '@coveo/atomic/components';
 import {createComponent} from '@lit/react';
 import React from 'react';
 
@@ -6,4 +9,10 @@ export const AtomicComponentError = createComponent({
   tagName: 'atomic-component-error',
   react: React,
   elementClass: LitAtomicComponentError,
+});
+
+export const AtomicIcon = createComponent({
+  tagName: 'atomic-icon',
+  react: React,
+  elementClass: LitAtomicIcon,
 });