+++ packages/atomic/src/components/common/item-list/item-decorators.tsx
@@ -1,6 +1,6 @@
 import {ComponentInterface, getElement} from '@stencil/core';
 import {buildCustomEvent} from '../../../utils/event-utils';
-import {closest} from '../../../utils/utils';
+import {closest} from '../../../utils/stencil-utils';
 import {AnyItem} from '../interface/item';
 import {
   ItemDisplayDensity,