+++ packages/atomic/src/components/search/atomic-search-interface/atomic-aria-live.tsx
@@ -9,7 +9,7 @@ import {
 } from '@stencil/core';
 import {buildDebouncedQueue} from '../../../utils/debounce-utils';
 import {FindAriaLiveEventArgs} from '../../../utils/stencil-accessibility-utils';
-import {randomID} from '../../../utils/utils';
+import {randomID} from '../../../utils/stencil-utils';
 
 type Regions = {[regionName: string]: {assertive: boolean; message: string}};