+++ packages/atomic/src/components/search/index.ts
@@ -1,3 +1,2 @@
-// Export the class of the components migrated to Lit here, like this:
-
-export {AtomicComponentError} from '../common/atomic-component-error/atomic-component-error';
+// Auto-generated file
+export {};