+++ packages/atomic/src/components/search/result-template-components/atomic-result-icon/e2e/atomic-result-icon.e2e.ts
@@ -1,0 +1,11 @@
+import {test, expect} from './fixture';
+
+test.describe('AtomicResultIcon', () => {
+  test.beforeEach(async ({resultIcon}) => {
+    await resultIcon.load();
+  });
+
+  test('atomic-result-icon loads an icon', async ({resultIcon}) => {
+    await expect(resultIcon.svg).toBeVisible();
+  });
+});