+++ packages/atomic/src/components/search/search-box-suggestions/atomic-search-box-recent-queries/atomic-search-box-recent-queries.tsx
@@ -6,7 +6,7 @@ import {
 import {Component, Element, Prop, State, h} from '@stencil/core';
 import Clock from '../../../../images/clock.svg';
 import {SafeStorage, StorageItems} from '../../../../utils/local-storage-utils';
-import {once} from '../../../../utils/utils';
+import {once} from '../../../../utils/stencil-utils';
 import {
   getPartialRecentQueryClearElement,
   getPartialRecentQueryElement,
@@ -20,7 +20,7 @@ import {
   SearchBoxSuggestionElement,
   SearchBoxSuggestions,
   SearchBoxSuggestionsBindings,
-} from '../../../common/suggestions/suggestions-common';
+} from '../../../common/suggestions/stencil-suggestions-common';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';
 
 /**